Moving tutorial state into page and drilling to components   - Copying game layout for initial implementation, will look at condensing and refactoring after   - Setup relative flex boxes for game layout so that modal doesn't fill entire page   - Created BlurSm and Highlight wrappers
