+++ yahtzee-front/components/tutorial/00_StartTutorial.tsx
