+++ yahtzee-front/components/tutorial/01_GameLayoutBlurWithModal.tsx
@@ -1,0 +1,32 @@
+
+import Highlight from "./Highlight"
+import Blur from "./BlurSm"
+import DiceBoard from "../game/DiceBoard"
+import GameFeed from "../game/GameFeed"
+import ScoreCard from "../game/ScoreCard"
+
+const GameLayoutBlurWithModal = ({scorecard, setScorecard, gameFeed, setGameFeed, user}) => {
+
+  return (
+    <div className="w-full h-full relative">
+      <Highlight>
+        <Blur>
+          <div className="absolute w-full h-full flex gap-10 p-5 border-4 border-orange-400 ">
+            <div className="w-1/2 h-full flex flex-col gap-5">
+              <DiceBoard scorecard={scorecard} setScorecard={setScorecard} gameFeed={gameFeed} setGameFeed={setGameFeed}/>
+              <GameFeed user={user} gameFeed={gameFeed} />
+            </div>
+            <div className="w-1/2 h-full ">
+              <ScoreCard scorecard={scorecard} />
+            </div>
+          </div>
+        </Blur>
+      </Highlight>
+      <div className="absolute w-full h-full top-0 left-0 flex items-center justify-center border-4 border-blue-400">
+        <div className="bg-white border-4 border-black text-black p-10 rounded-xl">This is the Dashboard</div>
+      </div>
+    </div>
+  )
+}
+
+export default GameLayoutBlurWithModal;
