+++ yahtzee-front/components/tutorial/02_GameLayout.tsx
@@ -1,0 +1,21 @@
+
+import DiceBoard from "../game/DiceBoard"
+import GameFeed from "../game/GameFeed"
+import ScoreCard from "../game/ScoreCard"
+
+const GameLayout = ({scorecard, setScorecard, gameFeed, setGameFeed, user}) => {
+
+  return (
+    <div className="w-full h-full absolute flex gap-10 p-5 ">
+      <div className="w-1/2 h-full flex flex-col gap-5">
+        <DiceBoard scorecard={scorecard} setScorecard={setScorecard} gameFeed={gameFeed} setGameFeed={setGameFeed}/>
+        <GameFeed user={user} gameFeed={gameFeed} />
+      </div>
+      <div className="w-1/2 h-full ">
+        <ScoreCard scorecard={scorecard} />
+      </div>
+    </div>
+  )
+}
+
+export default GameLayout;
